backend/requirements.txt
# FastAPI and related
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-decouple==3.8

# Database
tortoise-orm[asyncpg]==0.21.6
asyncpg==0.30.0
aerich==0.7.2

# GraphQL
strawberry-graphql[fastapi]==0.275.6

# AI/ML
tensorflow==2.18.0
pillow==11.0.0
numpy>=1.26.0,<2.1.0
opencv-python==4.10.0.84

# Monitoring
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0

# Caching
redis==5.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
boto3==1.35.92  # For S3/MinIO
aiofiles==24.1.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.1.1
httpx==0.28.1
structlog==24.4.0